+++ sqlparser/lib/src/ast/visitor.dart
@@ -17,6 +17,8 @@ abstract class AstVisitor<A, R> {
   R visitInvalidStatement(InvalidStatement e, A arg);
   R visitReindexStatement(ReindexStatement e, A arg);
   R visitVacuumStatement(VacuumStatement e, A arg);
+  R visitAttachStatement(AttachStatement e, A arg);
+  R visitDetachStatement(DetachStatement e, A arg);
 
   R visitReturning(Returning e, A arg);
   R visitWithClause(WithClause e, A arg);
@@ -180,6 +182,16 @@ class RecursiveVisitor<A, R> implements AstVisitor<A, R?> {
 
   @override
   R? visitVacuumStatement(VacuumStatement e, A arg) {
+    return visitStatement(e, arg);
+  }
+
+  @override
+  R? visitAttachStatement(AttachStatement e, A arg) {
+    return visitStatement(e, arg);
+  }
+
+  @override
+  R? visitDetachStatement(DetachStatement e, A arg) {
     return visitStatement(e, arg);
   }